+++ Apmtraces/requests/QueryRequest.cs
@@ -20,7 +20,7 @@ namespace Oci.ApmtracesService.Requests
     {
         
         /// <value>
-        /// The APM Domain ID the request is intended for.
+        /// The APM Domain ID for the intended request.
         /// 
         /// </value>
         /// <remarks>
@@ -80,7 +80,7 @@ namespace Oci.ApmtracesService.Requests
         public System.Nullable<int> Limit { get; set; }
         
         /// <value>
-        /// The page token representing the page at which to start retrieving results.
+        /// The page token representing the page where to start retrieving results.
         /// This is usually retrieved from a previous response.
         /// 
         /// </value>