+++ Common/Src/RegionDefinitions.cs
@@ -95,5 +95,8 @@ namespace Oci.Common
 
         // OC24
         public static readonly Region EU_DCC_ZURICH_1 = Register("eu-dcc-zurich-1", Realm.OC24, "avz");
+
+        // OC26
+        public static readonly Region ME_ABUDHABI_3 = Register("me-abudhabi-3", Realm.OC26, "ahu");
     }
 }